matchmaker/pyproject.toml
[project]
name = "matchmaker"
version = "0.1.0"
description = "Temporary Escape Matchmaker Server"
requires-python = ">=3.10"
authors = [
    { name = "Matus Novak", email = "email@matusnovak.com" }
]
dependencies = [
    "fastapi[all]",
    "sqlalchemy",
    "passlib[bcrypt]",
    "pydantic-settings",
    "python-jose[cryptography]",
    "pyinstaller",
    "black",
    "aiohttp",
    "ruff",
]

[tool.setuptools]
py-modules = []

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
